Show critical notes from mei elements in the web app.
we will be moving all notes from the current LaTEX files into
the MEI files. From that, we should generate the critical notes section
in the pdf output.

We should also experiment with the posibility of rendering the pdf in
different flavours for musicologiest, performers, etc... on there the
notes could be shown or not.