web-app: change how element with editorial items are shown.
Instead of puting boxes on our own overlay and track the original
position just change the original elements looks via CSS. Downside
is that notes bounding box are indeed one bb for the notehead and
a path for the stem, so filling the bb rect results in only the
notehead been. Even is we change the path color as we do in the
music playing tracking, we still have the problem of only the tiny
portion of the head getting pointer (and click events).

Maybe we should just highlight the entire measure and display
on the modal all the editorial items included on that measure