tono 34: Transcription and alto reconstruction fixes
This tono uses voice crossing as an aesthetic feature.  Sometimes it is
not clear if some other unusual patterns are also intentional or derived
from a transcription or scribe error.

So far I have changed the imitative approach in the alto reconstruction
and updated the coplas.  There are still many issues pending a proper
resolution