tono 48: rework alto section.
The first section of estribillo is indeed a question-answer between the Alto and the
rest of the voices, in a similar way the tenor does in the middle. Still
need to work out the alto solo line properly as this is currently a
placeholder to verify the whole structure. Also the section between m.55
and m.69 needs work as the texture is currently a mess.