tono 28: update the reconstructed alto in m28-34.
The texture in the CPMHL surviving parts is a fully layered entrance
while the surviving parts in CdM are two homophonic blocks. Try in the
reconstrudted alto an in-between solution. This info should be added to the
tono introductory reconstruction critical notes, as well as the guion
diferences betwen both versoins and our approach.